+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/sun/workflow/RenamingImageEmitter.java
@@ -42,32 +42,32 @@ public class RenamingImageEmitter implements ImageEmitter {
     }
 
     @Override
-    public void newMonoImage(GeneratedImageKind kind, String category, String title, String name, ImageWrapper32 image, Consumer<? super float[][]> bufferConsumer) {
-        delegate.newMonoImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), image, bufferConsumer);
+    public void newMonoImage(GeneratedImageKind kind, String category, String title, String name, String description, ImageWrapper32 image, Consumer<? super float[][]> bufferConsumer) {
+        delegate.newMonoImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), description, image, bufferConsumer);
     }
 
     @Override
-    public void newMonoImage(GeneratedImageKind kind, String category, String title, String name, ImageWrapper32 image) {
-        delegate.newMonoImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), image);
+    public void newMonoImage(GeneratedImageKind kind, String category, String title, String name, String description, ImageWrapper32 image) {
+        delegate.newMonoImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), description, image);
     }
 
     @Override
-    public void newColorImage(GeneratedImageKind kind, String category, String title, String name, ImageWrapper32 image, Function<ImageWrapper32, float[][][]> rgbSupplier) {
-        delegate.newColorImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), image, rgbSupplier);
+    public void newColorImage(GeneratedImageKind kind, String category, String title, String name, String description, ImageWrapper32 image, Function<ImageWrapper32, float[][][]> rgbSupplier) {
+        delegate.newColorImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), description, image, rgbSupplier);
     }
 
     @Override
-    public void newColorImage(GeneratedImageKind kind, String category, String title, String name, ImageWrapper32 image, Function<ImageWrapper32, float[][][]> rgbSupplier, BiConsumer<Graphics2D, ? super ImageWrapper> painter) {
-        delegate.newColorImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), image, rgbSupplier, painter);
+    public void newColorImage(GeneratedImageKind kind, String category, String title, String name, String description, ImageWrapper32 image, Function<ImageWrapper32, float[][][]> rgbSupplier, BiConsumer<Graphics2D, ? super ImageWrapper> painter) {
+        delegate.newColorImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), description, image, rgbSupplier, painter);
     }
 
     @Override
-    public void newColorImage(GeneratedImageKind kind, String category, String title, String name, int width, int height, Map<Class<?>, Object> metadata, Supplier<float[][][]> rgbSupplier) {
-        delegate.newColorImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), width, height, metadata, rgbSupplier);
+    public void newColorImage(GeneratedImageKind kind, String category, String title, String name, String description, int width, int height, Map<Class<?>, Object> metadata, Supplier<float[][][]> rgbSupplier) {
+        delegate.newColorImage(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), description, width, height, metadata, rgbSupplier);
     }
 
     @Override
-    public void newGenericFile(GeneratedImageKind kind, String category, String title, String name, Path file) {
-        delegate.newGenericFile(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), file);
+    public void newGenericFile(GeneratedImageKind kind, String category, String title, String name, String description, Path file) {
+        delegate.newGenericFile(kind, null, titleRenamer.apply(title), fileRenamer.apply(name), description, file);
     }
 }